refactor: split projects.$projectId.tsx into domain components
Split 516-line monolithic route into focused components:

- parse-env.ts: Pure utility for .env parsing with discriminated union result type
- ProjectHeader.tsx: Project metadata and secrets toggle (~43 lines)
- SecretsPanel.tsx: Secrets CRUD, bulk import, and editing (~316 lines)
- SessionsGrid.tsx: Sessions list with create/resume actions (~137 lines)
- projects.$projectId.tsx: Thin composition layer (~58 lines, down from 516)

Type-safe throughout, mutations colocated with their UI state.
